+++ src/biz/bokhorst/xprivacy/Meta.java
@@ -230,6 +230,18 @@ public class Meta {
 
 		mListHook.add(new Hook("sensors", "getDefaultSensor", "", 3, null, null));
 		mListHook.add(new Hook("sensors", "getSensorList", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "acceleration", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "gravity", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "humidity", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "light", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "magnetic", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "motion", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "orientation", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "pressure", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "proximity", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "rotation", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "temperature", "", 3, null, null));
+		mListHook.add(new Hook("sensors", "step", "", 3, null, null));
 
 		mListHook.add(new Hook("shell", "sh", "", 10, "0.0", null).dangerous().whitelist(cTypeCommand));
 		mListHook.add(new Hook("shell", "su", "", 10, "0.0", null).dangerous().whitelist(cTypeCommand));