+++ src/biz/bokhorst/xprivacy/XSensorManager.java
@@ -46,17 +46,73 @@ public class XSensorManager extends XHook {
 		if (mMethod == Methods.getDefaultSensor) {
 			if (isRestricted(param))
 				param.setResult(null);
+			else if (param.args.length > 0)
+				if (isRestricted(param, (Integer) param.args[0]))
+					param.setResult(null);
 
 		} else if (mMethod == Methods.getSensorList) {
 			if (isRestricted(param))
 				param.setResult(new ArrayList<Sensor>());
+			else if (param.args.length > 0)
+				if (isRestricted(param, (Integer) param.args[0]))
+					param.setResult(new ArrayList<Sensor>());
 
 		} else
 			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
 	}
 
 	@Override
+	@SuppressWarnings("unchecked")
 	protected void after(XParam param) throws Throwable {
-		// Do nothing
+		if (mMethod == Methods.getSensorList)
+			if (param.getResult() != null && param.args.length > 0 && (Integer) param.args[0] == Sensor.TYPE_ALL) {
+				List<Sensor> listSensor = new ArrayList<Sensor>();
+				for (Sensor sensor : (List<Sensor>) param.getResult())
+					if (!isRestricted(param, sensor.getType()))
+						listSensor.add(sensor);
+				param.setResult(listSensor);
+			}
+	}
+
+	private boolean isRestricted(XParam param, int type) throws Throwable {
+		if (type == Sensor.TYPE_ACCELEROMETER || type == Sensor.TYPE_LINEAR_ACCELERATION) {
+			if (isRestricted(param, "acceleration"))
+				return true;
+		} else if (type == Sensor.TYPE_GRAVITY) {
+			if (isRestricted(param, "gravity"))
+				return true;
+		} else if (type == Sensor.TYPE_RELATIVE_HUMIDITY) {
+			if (isRestricted(param, "humidity"))
+				return true;
+		} else if (type == Sensor.TYPE_LIGHT) {
+			if (isRestricted(param, "light"))
+				return true;
+		} else if (type == Sensor.TYPE_MAGNETIC_FIELD || type == Sensor.TYPE_MAGNETIC_FIELD_UNCALIBRATED) {
+			if (isRestricted(param, "magnetic"))
+				return true;
+		} else if (type == Sensor.TYPE_SIGNIFICANT_MOTION) {
+			if (isRestricted(param, "motion"))
+				return true;
+		} else if (type == Sensor.TYPE_GYROSCOPE || type == Sensor.TYPE_GYROSCOPE_UNCALIBRATED) {
+			if (isRestricted(param, "orientation"))
+				return true;
+		} else if (type == Sensor.TYPE_PRESSURE) {
+			if (isRestricted(param, "pressure"))
+				return true;
+		} else if (type == Sensor.TYPE_PROXIMITY) {
+			if (isRestricted(param, "proximity"))
+				return true;
+		} else if (type == Sensor.TYPE_GAME_ROTATION_VECTOR || type == Sensor.TYPE_GEOMAGNETIC_ROTATION_VECTOR
+				|| type == Sensor.TYPE_ROTATION_VECTOR) {
+			if (isRestricted(param, "rotation"))
+				return true;
+		} else if (type == Sensor.TYPE_AMBIENT_TEMPERATURE) {
+			if (isRestricted(param, "temperature"))
+				return true;
+		} else if (type == Sensor.TYPE_STEP_COUNTER || type == Sensor.TYPE_STEP_DETECTOR) {
+			if (isRestricted(param, "step"))
+				return true;
+		}
+		return false;
 	}
 }